move index interfaces
